Moved dtype tests to appropriate file + improved docs
